Move make_service and make_issue helpers to tests.base
